requirements.txt
asgiref==3.10.0
Django==4.2
djangorestframework==3.16.1
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
